mp7/requirements.txt
pandas
requests
python-dotenv
pytest
flask
